requirements.txt
# pip install -r requestments.txt
PyYAML==5.4.1
Shapely==1.8.rc1
